+++ src/context/colorModeContext.ts
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const ColorModeContext = createContext({ toggleColorMode: () => {} });
